Add DebugWritableStream to LogInterceptor
